modify: modify route & routes
