+++ lib/router/router.js
@@ -1,14 +1,64 @@
-import { renderAll } from '../render/render.js';
-import { navigate } from './navigate.js';
+// pages
+import { pageLogIn } from '../../src/pages/login/sign_in.js';
+import { pageSignUp } from '../../src/pages/login/sign_up.js';
+import { pageProfile } from '../../src/pages/profile.js';
+import { pageGame } from '../../src/pages/game.js';
+import { pageTournament } from '../../src/pages/tournament.js';
+import { pageSwitch } from '../../src/pages/switch.js';
+import { pageTwoFA } from '../../src/pages/login/twofa.js';
+import { PageNotFound } from '../../src/pages/404.js';
+// state
 import {
   gameState,
   globalState,
   routeState,
   userState,
 } from '../state/state.js';
+// render
+import { renderAll } from '../render/render.js';
+// router
+import { navigate } from './navigate.js';
+// utils
 import { initUserInfo } from '../../src/utils/initUser.js';
 
-export function routeByState(href = window.location.pathname) {
+export const routes = {
+  '/login': { name: 'Login', page: pageLogIn, component: [], onRender: [] },
+  '/signup': {
+    name: 'Signup',
+    page: pageSignUp,
+    component: [],
+    onRender: [],
+  },
+  '/profile': {
+    name: 'Profile',
+    page: pageProfile,
+    component: [],
+    onRender: [],
+  },
+  '/game': { name: 'Game', page: pageGame, component: [], onRender: [] },
+  '/tournament': {
+    name: 'Tournament',
+    page: pageTournament,
+    component: [],
+    onRender: [],
+  },
+  '/logout': {
+    name: 'Logout',
+    page: pageSwitch,
+    component: [],
+    onRender: [],
+  },
+  '/twofa': { name: 'TwoFA', page: pageTwoFA, component: [], onRender: [] },
+  '/switch': {
+    name: 'Switch',
+    page: pageSwitch,
+    component: [],
+    onRender: [],
+  },
+  '/404': { name: '404', page: PageNotFound, component: [], onRender: [] },
+};
+
+export function routeByState(href) {
   routeState.ping();
 
   if (!globalState.getState().isLoggedIn) {
@@ -21,18 +71,25 @@ export function routeByState(href = window.location.pathname) {
 }
 
 // 기본 경로를 설정. (/ -> /profile)
-export function setDefaultPath(href, routes) {
+export function setDefaultPath(href = window.location.pathname) {
   routes['/'] = routes['/profile']; // navbar에 경로 추가가 되지 않도록 함
 
-  return routeByState(href);
+  if (href === '/' || href === '/login') {
+    return '/';
+  } else {
+    return href;
+  }
 }
 
-export function redirectRoute(routes, path) {
+export function redirectRoute(path) {
   const target = navigate(routes, path, true);
   renderAll(target);
 
   routeState.setState({ currentRoute: target }, false); // 무한 루프 방지
-  userState.ping(); // userBox 업데이트
+
+  if (globalState.getState().isLoggedIn) {
+    userState.ping(); // userBox 업데이트
+  }
 
   if (target.onRender) {
     target.onRender.forEach((func) => {
@@ -44,7 +101,6 @@ export function redirectRoute(routes, path) {
 let previousPath = '';
 
 export function route(
-  routes,
   path = window.location.pathname,
   updateHistory = true,
   callListeners = true
@@ -75,26 +131,23 @@ export function route(
   }
 }
 
-export function firstRoute(routes, path) {
+export function firstRoute(path) {
   const target = navigate(routes, path, false);
   renderAll(target);
 
-  routeState.setState({ currentRoute: target }, false);
+  routeState.setState({ currentRoute: target });
+  gameState.setState({ currentGameStatus: 'idle' });
 
   if (globalState.getState().isLoggedIn) {
     initUserInfo().then(() => {
       userState.ping(); // userBox 업데이트
-      if (target.onRender) {
-        target.onRender.forEach((func) => {
-          func();
-        });
-      }
     });
-  } else {
+
     if (target.onRender) {
       target.onRender.forEach((func) => {
         func();
       });
     }
   }
+  console.log(path, target);
 }
