+++ src/components/login/sign_in.js
@@ -1,6 +1,5 @@
 import { route } from '../../../lib/router/router.js';
-import { routes } from '../../app.js';
-import { userState } from '../../../lib/state/state.js';
+import { globalState, userState } from '../../../lib/state/state.js';
 import { setCookie } from '../../../src/utils/cookie.js';
 
 // [로그인 요청]
@@ -22,7 +21,7 @@ async function requestLogin(credentials) {
       const responseData = await response.json(); // 비동기
       setCookie(responseData);
 
-      route(routes, '/profile', true, false);
+      route('/profile', true, false);
     } else {
       throw new Error(response.status.toString());
     }
