+++ src/components/login/sign_up.js
@@ -1,6 +1,5 @@
 import { userState } from '../../../lib/state/state.js';
 import { route } from '../../../lib/router/router.js';
-import { routes } from '../../app.js';
 import { setCookie } from '../../../src/utils/cookie.js';
 import { validateUsername, validatePassword } from '../../utils/validator.js';
 
@@ -26,7 +25,7 @@ async function requestSignUp(formData) {
     });
     setCookie(data);
 
-    route(routes, '/twofa', true, false);
+    route('/twofa', true, false);
     return data;
   } catch (e) {
     console.log(e);
